requirements.txt
discord.py
socketio